Implement function to read a configuration from a specified group in an h5 file, for example restart files.
This function is probably is only temporary until Lorenzo implements a more general and unified system for reading configurations. It takes a group name, for example restarts/restart0000 and reads the configuration including the correct simulation box (Orthorhombic or LeesEdwards) which the existing configuration_from_hdf5 function does not do.