Implement scale function for simulation_boxes. Remove lengths as a member.
Box lengths can be scaled using the new function scale which works both with a single float and an array of length D. Later it could be a matrix. Access to the lengths is now through the method get_lengths. Tests and examples have been updated. Tinkering with lengths to account for walls etc, as seen in the (currently not working) example LJchain_wall.py can be done by creating a separate array with the desired lengths and then applying scale to the ratio of newlengths to old.